requirements.txt
django
django-rest-framework-mongoengine
rest_framework